Issues-19: Moved example of com.github.iammert:ScalingLayout to separate module.
