+++ app/src/main/java/com/kirchhoff/animation/MainActivity.kt
@@ -4,15 +4,14 @@ import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.View
-
 import com.example.TickerActivity
 import com.example.alerter.AlerterActivity
 import com.example.loadingbutton.LoadingButtonActivity
 import com.kirchhoff.animation.expandable.ExpandableLinearLayoutActivity
 import com.kirchhoff.animation.recycler.focusresize.FocusResizeActivity
 import com.kirchhoff.animation.recycler.vega.VegaManagerActivity
-import com.kirchhoff.animation.scaling.ScalingMainActivity
 import com.kirchhoff.particle.ParticleActivity
+import com.kirchhoff.scalinglayout.ScalingMainActivity
 import com.kirchhoff.swipeback.SwipeBackLayoutActivity
 
 @Suppress("UNUSED_PARAMETER")
