+++ app/src/main/res/layout/activity_main.xml
@@ -76,7 +76,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="scalingLayoutExample"
-            android:text="@string/scaling.layout" />
+            android:text="@string/scaling_layout" />
 
     </LinearLayout>
