+++ scalinglayout/src/main/res/drawable/ic_date_range.xml
