+++ scalinglayout/src/main/res/drawable/ic_star.xml
