+++ scalinglayout/src/main/res/drawable/ic_watch_later.xml
