+++ scalinglayout/src/main/res/layout/a_scaling_fab.xml
