+++ scalinglayout/src/main/res/layout/a_scaling_search_bar.xml
